Move things in pipeline module
