+++ myia/pipeline/__init__.py
@@ -1,0 +1,17 @@
+"""Contains Myia's pipeline definitions, steps, resources, etc."""
+
+
+from .pipeline import (  # noqa
+    PipelineDefinition, Pipeline, PipelineResource, PipelineStep,
+    pipeline_function
+)
+
+from .standard import (  # noqa
+    standard_resources,
+    standard_pipeline,
+    standard_debug_pipeline,
+    scalar_pipeline,
+    scalar_debug_pipeline,
+    scalar_parse,
+    scalar_debug_compile
+)
