+++ tests/debug/test_debug_utils.py
@@ -3,7 +3,7 @@
 
 import pytest
 
-from myia.api import scalar_parse as parse
+from myia.pipeline import scalar_parse as parse
 from myia.debug.utils import GraphIndex
