+++ tests/test_specialize.py
@@ -2,7 +2,7 @@
 import numpy
 from pytest import mark
 
-from myia.api import scalar_debug_pipeline, standard_debug_pipeline
+from myia.pipeline import scalar_debug_pipeline, standard_debug_pipeline
 from myia.composite import list_map
 from myia.debug.label import short_labeler as lbl
 from myia.debug.traceback import print_inference_error
